+++ r4/src/test/java/io/github/jy95/fds/r4/translators/BoundsRangeTest.java
@@ -1,31 +1,25 @@
 package io.github.jy95.fds.r4.translators;
 
+import io.github.jy95.fds.common.types.DosageAPI;
 import io.github.jy95.fds.r4.DosageAPIR4;
 import io.github.jy95.fds.r4.AbstractFhirTest;
 import io.github.jy95.fds.common.types.DisplayOrder;
+import io.github.jy95.fds.r4.config.FDSConfigR4;
+import io.github.jy95.fds.translators.AbstractBoundsRangeTest;
 import org.hl7.fhir.r4.model.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.ExecutionException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class BoundsRangeTest extends AbstractFhirTest {
+public class BoundsRangeTest extends AbstractBoundsRangeTest<FDSConfigR4, Dosage> {
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testNoBoundsRange(Locale locale) throws ExecutionException, InterruptedException {
-        Dosage dosage = new Dosage();
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        assertEquals("", result);
-    }
-
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testBothLowAndHighWithoutUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateBothLowAndHighWithoutUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -35,15 +29,11 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText1(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testBothLowAndHighWithUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateBothLowAndHighWithUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -53,15 +43,11 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText2(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyHighWithoutUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyHighWithoutUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -70,15 +56,11 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText3(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyHighWithUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyHighWithUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -87,15 +69,11 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText4(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyLowWithoutUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyLowWithoutUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -104,15 +82,11 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText5(locale);
-        assertEquals(expected, result);
+        return dosage;
     }
 
-    @ParameterizedTest
-    @MethodSource("localeProvider")
-    void testOnlyLowWithUnit(Locale locale) throws ExecutionException, InterruptedException {
+    @Override
+    protected Dosage generateOnlyLowWithUnit() {
         Dosage dosage = new Dosage();
         Timing timing = new Timing();
         Timing.TimingRepeatComponent timingRepeatComponent = new Timing.TimingRepeatComponent();
@@ -121,87 +95,24 @@ public class BoundsRangeTest extends AbstractFhirTest {
         timingRepeatComponent.setBounds(boundsRange);
         timing.setRepeat(timingRepeatComponent);
         dosage.setTiming(timing);
-        DosageAPIR4 dosageUtils = getDosageAPI(locale, DisplayOrder.BOUNDS_RANGE);
-        String result = dosageUtils.asHumanReadableText(dosage).get();
-        String expected = getExpectedText6(locale);
-        assertEquals(expected, result);
+        return dosage;
+    }
+    
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(Locale locale, DisplayOrder displayOrder) {
+        return new DosageAPIR4(FDSConfigR4.builder()
+                .displayOrder(List.of(displayOrder))
+                .locale(locale)
+                .build());
     }
 
-    // For the parametrized test of first test
-    private static String getExpectedText1(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for 1 to 3";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour 1 à 3";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für zwischen 1 und 3";
-        } else {
-            return "gedurende tussen 1 en 3";
-        }
+    @Override
+    public DosageAPI<FDSConfigR4, Dosage> getDosageAPI(FDSConfigR4 config) {
+        return new DosageAPIR4(config);
     }
 
-    // For the parametrized test of second test
-    private String getExpectedText2(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for 1 to 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour 1 à 3 jours";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für zwischen 1 und 3 Tage";
-        } else {
-            return "gedurende tussen 1 en 3 dagen";
-        }
-    }
-
-    // For the parametrized test of third test
-    private String getExpectedText3(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for up to 3";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour 3 maximum";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für bis 3";
-        } else {
-            return "gedurende tot 3";
-        }
-    }
-
-    // For the parametrized test of four test
-    private String getExpectedText4(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for up to 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour 3 jours maximum";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für bis 3 Tage";
-        } else {
-            return "gedurende tot 3 dagen";
-        }
-    }
-
-    // For the parametrized test of fifth test
-    private String getExpectedText5(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for at least 3";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour au moins 3";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für mindestens 3";
-        } else {
-            return "gedurende minstens 3";
-        }
-    }
-
-    // For the parametrized test of sixth test
-    private String getExpectedText6(Locale locale) {
-        if (locale.equals(Locale.ENGLISH)) {
-            return "for at least 3 days";
-        } else if (locale.equals(Locale.FRENCH)) {
-            return "pour au moins 3 jours";
-        } else if (locale.equals(Locale.GERMAN)) {
-            return "für mindestens 3 Tage";
-        } else {
-            return "gedurende minstens 3 dagen";
-        }
+    @Override
+    public Dosage generateEmptyDosage() {
+        return new Dosage();
     }
 }
